Time: 1631 ms (22.43%) | Memory: 33.1 MB (90.85%) - LeetSync

+++ 560-subarray-sum-equals-k/subarray-sum-equals-k.cpp
@@ -3,7 +3,7 @@ public:
     int subarraySum(vector<int>& nums, int k)
     {
         int n = nums.size();
-        int ans = 0;
+        double ans = 0;
         for( int  i =0; i < n; i++)
         {
             int sum = nums[i];
